Moved text from Text element infopage to Typograhy.
MIMIR-276
