+++ src/pages/Components/componentsList.jsx
@@ -51,8 +51,6 @@ import TabsInfo from './ComponentsInfo/TabsInfo';
 import tabsIcon from '../../../public/customIcons/tabs.svg';
 import TagsInfo from './ComponentsInfo/TagsInfo';
 import tagsIcon from '../../../public/customIcons/tags.svg';
-import TypographyInfo from './ComponentsInfo/TypographyInfo';
-import typographyIcon from '../../../public/customIcons/typography.svg';
 
 const componentsList = [
 	{ component: AccordionInfo, label: 'Accordion', path: '/accordion', icon: accordionIcon },
@@ -84,7 +82,6 @@ const componentsList = [
 	{ component: '', label: 'Table', path: '/table', icon: tableIcon, comingSoon: true },
 	{ component: TabsInfo, label: 'Tabs', path: '/tabs', icon: tabsIcon },
 	{ component: TagsInfo, label: 'Tags', path: '/tags', icon: tagsIcon },
-	{ component: TypographyInfo, label: 'Text elements', path: '/text-elements', icon: typographyIcon },
 	{ component: WarningsErrorsInfo, label: 'Warnings and errors', path: '/warnings-errors', icon: dialogIcon },
 ];
